### 2016-09-21 #77 @whedon
[PRE REVIEW]: Habfuzz: A Fortran tool to calculate the instream hydraulic habitat suitability based on fuzzy logic
**Submitting author:** @chtheodoro (<a href="http://orcid.org/0000-0002-5395-0347">Christos Theodoropoulos</a>)
**Repository:** <a href="https://github.com/chtheodoro/Habfuzz" target ="_blank">https://github.com/chtheodoro/Habfuzz</a>
**Version:** v1.0.0
