### 2016-09-23 a61b22b
Minor amendments
