+++ README.md
@@ -116,7 +116,7 @@ After having the input files ready, double click habfuzz.exe. The command prompt
 
 There are three available scenarios based on the method used for deriving the outcome of each IF-THEN rule from the reference conditions of the program, (i) the moderate scenario, where the different suitability values for the same combinations of flow velocity, water depth and substrate type are averaged to derive the final suitability, (ii) the worst scenario, where the final suitability is derived from the minimum observed suitability and (iii) the optimum scenario where the final suitability is derived by the maximum observed suitability. A default scenario is also present (the moderate scenario). Note that if a specific combination in the observed data does not match a combination in the reference data, the program returns a value of ‘-1’ for the habitat suitability.    
 
-After selecting the desired scenario, the user is asked to select the defuzzification method (see section 1). A default method (centroid) is also available. After selecting the defuzzification method, Habfuzz calls the relevant FORTRAN subroutines to perform the tasks selected. The program informs the user when the process is completed and indicates the suitability.txt file created where the suitability values are stored. The file is located in the ‘habfuzz’ subfolder.
+After selecting the desired scenario, the user is asked to select the defuzzification method (see section 1). A default method (centroid) is also available. After selecting the defuzzification method, Habfuzz calls the relevant Fortran subroutines to perform the tasks selected. The program informs the user when the process is completed and indicates the suitability.txt file created where the suitability values are stored. The file is located in the ‘habfuzz’ subfolder.
 
 ![9](https://cloud.githubusercontent.com/assets/21544603/18738401/027f675a-80a4-11e6-8608-10ebeca67071.png)
 
@@ -130,11 +130,11 @@ a. FUZZIFICATION - The fuzzification process of Habfuzz is included in the subro
 
 ![fig5](https://cloud.githubusercontent.com/assets/21544603/18668781/325aad78-7f40-11e6-90c8-acf60eaac17b.png)
 
-c. IF-THEN RULES - The user can also apply modifications to the IF-THEN rules of the smod.f95, swors.f95 and sopt.f95 subroutines. As already mentioned, the reference conditions are currently derived for freshwater macroinvertebrates, sampled in extended sampling campaign, which took place in 9 sampling sites in Greece and resulted in a set of 380 reference microhabitats (combinations of flow velocity, water depth and substrate type). If the user needs to apply his/her own reference data (probably for other biotic elements of the aquatic ecosystem or even in a totally different application), changes should be applied to one of the abovementioned subroutines. However, if for example, the smod.f95 file is modified, the user should also change the column number of the arrays, which are used by smod.f95 (in B. INTERNAL PARAMETERS TO FACILITATE THE IMPLICATION AND AGGREGATION STEPS, marked as ‘Used by smod.f95’), which are the arrays g, h, m, p, b. The modification should be such that the column number is the same as the times that a combination in the smod.f95 file results to a specific result. If you see for example, ‘g’ is the outcome of 26 combinations and so the column number for ‘g’ in the fdeclarations.f95 file is also 26.
[…]
