### 2016-09-25 #77 @labarba
@fonnesbeck — I'm a bit at a loss here to shepherd the review of this paper, as I'm unfamiliar with the methods and application mentioned in the description ... I remember that you worked in some ecological applications of stats and data science: would you be able to review this paper for JOSS, or to suggest another appropriate reviewer?
Reviewer guidelines at:
http://joss.theoj.org/about#reviewer_guidelines
