### 2016-09-25 #77 @chtheodoro
If I am allowed to propose, maybe @jsta could help here as I see he has a strong ecological background. Although we need someone familiar with fuzzy logic, the submission is not hard to follow if you see the README.md. I am here also for any questions 😊
