### 2016-09-25 #77 @fonnesbeck
@labarba I will have a look, and either review it myself or suggest another. I have done some stream ecology work in the past, but not these methods.
