### 2016-09-27 #82 @whedon
[REVIEW]: Habfuzz: A Fortran tool to calculate the instream hydraulic habitat suitability based on fuzzy logic
**Submitting author:** @chtheodoro (<a href="http://orcid.org/0000-0002-5395-0347">Christos Theodoropoulos</a>)
**Repository:** <a href="https://github.com/chtheodoro/Habfuzz" target ="_blank">https://github.com/chtheodoro/Habfuzz</a>
**Version:** v1.0.0
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.163291" target="_blank">https://dx.doi.org/10.5281/zenodo.163291</a>
