### 2016-09-27 #77 @arfon
OK, the review is issue is over here: https://github.com/openjournals/joss-reviews/issues/82

@fonnesbeck - please take a look at that issue for completing the review.
