### 2016-09-27 ffcb767
Update paper.md
