### 2016-09-27 91dcf96
Update paper.md
