### 2016-09-27 d88b522
Update paper.md
