+++ paper/paper.bib
@@ -13,6 +13,7 @@
     journal = {International Journal of Man-Machine Studies},
     volume = {7},
     pages = {1--13},
+    DOI = 10.1016/S0020-7373(75)80002-2
 }
 
 @book{fuzzy,
