+++ paper/paper.bib
@@ -25,9 +25,9 @@
 }
 
 @online{github,
-    author = {Theodoropoulos Christos and Skoulikidis Nikolaos and Stamou Anastasios},
+    author = {Theodoropoulos C. and Skoulikidis N. and Stamou A.},
     title = {Habfuzz: A Fortran tool to calculate the instream hydraulic habitat suitability based on fuzzy logic},
     year = 2016,
-    url = {https://github.com/chtheodoro/Habfuzz},
+    url = {https://github.com/chtheodoro/habfuzz},
     urldate = {2016-09-21}
 }
