### 2016-10-04 #82 @fonnesbeck
Unless I am missing something, there is no clear specification of the platforms that the software is compatible with. It appears that it is MS-Windows-only, based on the fact that the installation instructions tell the user to "double-click habfuzz.exe to run the program".

If this is the case, I will not be able to review the paper, as I have no easy access to a Windows machine. It would be nice if the software were able to run on some sort of POSIX platform, but at a minimum the platform compatibilities need to be stated explicitly in the installation instructions, if not earlier.
