### 2016-10-05 #82 @chtheodoro
'Habfuzz has been developed to run on 32-bit and 64-bit Windows operating systems, including Windows XP, Vista, 7/8 and 10'. I have now inserted the above text at the 'instalation' section.
