### 2016-10-05 #82 @labarba
I am disinclined to accept in JOSS a piece of software that is Windows only. Although we (the JOSS editors) decided that open-source packages that require a proprietary runtime would be acceptable (reluctantly), that was a decision aimed at things like Matlab code. Software restricting the OS to Windows is in my opinion incompatible with the ethos of open source.
