### 2016-10-05 #82 @chtheodoro
I must admit I am not a software developer by profession, I am using Windows and developed this program in Windows. However, considering that the source files are read in every OS I suppose that if someone compiles the fortan files with a Linux-compatible Fortran compiler, it should work. If I am wrong and runs only in Windows then it's ok if you can't accept it.
