### 2016-10-05 #82 @fonnesbeck
The gfortran commands in the `.bat` file seemed to compile just fine, but it looks like the user interface would be the piece that would require some work, if I am interpreting this correctly.
