### 2016-10-05 #82 @chtheodoro
Habfuzz comes without a GUI. It inputs 3 .txt files and outputs 1 .txt file with the suitability in an array. I thought the lack of a GUI is not a problem...  is it?
