### 2016-10-05 #82 @fonnesbeck
I assumed the .exe opened some sort of interface (not necessarily graphical), based on the documentation. No, there is no GUI requirement.
