### 2016-10-05 #82 @chtheodoro
Thank you @fonnesbeck. Please let me know if the review is still ongoing in order to update the README.md instructions. Sorry for any inconvenience caused...
