### 2016-10-05 #82 @fonnesbeck
I am not able to review this, as I do no not readily have access to a Windows machine, either at home or at work. Perhaps we can find someone who does.
