### 2016-10-05 #82 @chtheodoro
I was wondering if instead of not accepting, I could have some time to test the program and make it run in Ubuntu and Mac OS, with simultaneously updating the installation Information in the README.md...
