+++ habfuzz/velocities.txt
@@ -1,2 +1,11 @@
-1
-0.08
+10
+0.28
+0.29
+0.01
+0.19
+0.68
+1.15
+0.89
+0.07
+0.31
+0.23
