+++ habfuzz/suitability.txt
@@ -1,0 +1,10 @@
+  0.671
+  0.440
+  0.690
+  0.737
+  0.300
+ -1.000
+ -1.000
+  0.367
+  0.480
+  0.500
