+++ habfuzz/log.txt
@@ -1,0 +1,179 @@
+  V-Very high membership class
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  1.000
+  0.000
+  0.000
+  0.000
+  0.000
+  
+  V-High membership class
+  0.000
+  0.000
+  0.000
+  0.000
+  1.000
+  0.000
+  1.000
+  0.000
+  0.000
+  0.000
+  
+  V-Moderate membership class
+  0.200
+  0.400
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.800
+  0.000
+  
+  V-Low membership class
+  0.800
+  0.600
+  0.000
+  1.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.200
+  1.000
+  
+  V-Very low membership class
+  0.000
+  0.000
+  1.000
+  0.000
+  0.000
+  0.000
+  0.000
+  1.000
+  0.000
+  0.000
+  
+  D-Very deep membership class
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  
+  D-Deep membership class
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  1.000
+  
+  D-Moderate membership class
+  1.000
+  1.000
+  0.000
+  0.000
+  1.000
+  0.000
+  0.000
+  0.000
+  1.000
+  0.000
+  
+  D-Shallow membership class
+  0.000
+  0.000
+  1.000
+  1.000
+  0.000
+  0.000
+  0.000
+  0.667
+  0.000
+  0.000
+  
+  D-Very shallow membership class
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  1.000
+  1.000
+  0.333
+  0.000
+  0.000
+  
+  Joint probability of High suitability class
+  0.295
+  0.000
+  0.280
+  0.546
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  
+  Joint probability of Good suitability class
+  0.324
+  0.000
+  0.440
+  0.274
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  0.000
+  
+  Joint probability of Moderate suitability class
+  0.324
+  0.700
+  0.230
+  0.090
+  0.000
+  0.000
+  0.000
+  0.667
+  0.900
+  1.000
+  
+  Joint probability of Poor suitability class
+  0.057
+  0.300
+  0.050
+  0.000
+  1.000
+  0.000
+  0.000
+  0.000
+  0.100
+  0.000
+  
+  Joint probability of Bad suitability class
+  0.000
+  0.000
+  0.000
+  0.090
+  0.000
+  0.000
+  0.000
+  0.333
+  0.000
+  0.000
