+++ habfuzz/substrates.txt
@@ -1,2 +1,11 @@
-1
+10
+0.040
+0.030
+0.070
+0.070
+0.030
+0.050
+0.070
 0.022
+0.030
+0.070
