+++ habfuzz/depths.txt
@@ -1,2 +1,11 @@
-1
-0.06
+10
+0.42
+0.37
+0.27
+0.17
+0.55
+0.04
+0.05
+0.09
+0.54
+0.76
