### 2016-10-07 51b4e0d
Update paper.md
