+++ paper/paper.bib
@@ -1,3 +1,12 @@
+@conference{bayes,
+    author = {Brooks C.J. and Kumar V. and Lane S.N.},
+    title = {A comparison of Fuzzy, Bayesian and Weighted Average formulations of an in-stream habitat suitability model},
+    conference = {International Congress on Environmental Modelling and Software},
+    year = {2010},
+    url = {http://www.iemss.org/iemss2010/papers/S20/S.20.07.Model%20selection%20and%20uncertainty%20A%20comparison%20of%20Fuzzy,%20Bayesian%20and%20Weighted%20Average%20formulations%20of%20an%20instream%20habitat%20suitability%20model%20-%20CHRISTOPHER%20BROOKES.pdf},
+    urldate = {2016-09-21},
+}
+
 @book{Fortran,
     author = {Adams J.C. and Brainerd W.S. and Martin J.T. and Smith B.T. and Wagener J.L.},
     title = {Fortran 95 Handbook - Complete ISO/ANSI Reference},
