+++ habfuzz/suitability.txt
@@ -1,10 +1,3000 @@
-  0.671
-  0.440
-  0.690
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.705
+ -1.000
+ -1.000
+  0.717
+  0.900
+  0.900
+ -1.000
+ -1.000
+ -1.000
+  0.900
+  0.717
+ -1.000
+ -1.000
+  0.717
+  0.900
+  0.900
+ -1.000
+ -1.000
+ -1.000
+  0.900
+  0.717
+ -1.000
+ -1.000
+  0.717
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.900
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.717
+  0.837
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.717
+ -1.000
+ -1.000
+  0.717
+  0.757
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.717
+  0.900
+  0.900
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.679
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.900
+  0.675
+ -1.000
+ -1.000
+  0.675
+  0.700
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.716
+  0.900
+  0.700
+ -1.000
+ -1.000
+ -1.000
+  0.700
+  0.675
+ -1.000
+ -1.000
+  0.717
+  0.875
+  0.700
[…]
