### 2016-10-12 b41612e
Update paper.md
