+++ paper/paper.md
@@ -20,7 +20,9 @@ bibliography: paper.bib
 ---
 
 # Summary
-Habfuzz is a Fortran tool, which implements the Mamdani - Assilian fuzzy inference process (Mamdani and Assilian, 1975) and the Bayesian joint probability inference method as described in Brookes et al. (2010), but with fuzzified inputs in order to calculate the instream habitat suitability, given the flow velocity, the water depth and the substrate type of a hydraulically simulated river reach. While it is specifically structured to quickly calculate the fuzzy-logic- (or fuzzy-Bayesian)-based instream habitat suitability for fish or freshwater macroinvertebrates, it can be applied to wider research topics requiring fuzzy logic to be addressed. The main Habfuzz routine reads in three files containing the flow velocities (V), the water depths (D) and the substrate types (S), as derived from a hydrodynamic/hydraulic simulation of the river reach (each node of the computational grid being an array element) and three additional user-defined parameters (the inference process to be implemented, the management scenario and the defuzzification method) and calls the relevant subroutines to implement the selected inference process. The output of Habfuzz is a .txt file containing an array with the habitat suitabilities for each node of the initial computational grid of the hydrodynamic simulation and a .txt log file containing the internal parameters of the inference process.
+Habfuzz is a Fortran tool, which implements the Mamdani - Assilian fuzzy inference process (Mamdani and Assilian, 1975) and the Bayesian joint probability inference method as described in Brookes et al. (2010), but with fuzzified inputs in order to calculate the instream habitat suitability, given the flow velocity, the water depth and the substrate type of a hydraulically simulated river reach. While it is specifically structured to quickly calculate the fuzzy-logic- (or fuzzy-Bayesian)-based instream habitat suitability for fish or freshwater macroinvertebrates, it can be applied to wider research topics requiring fuzzy logic to be addressed.
+
+The main Habfuzz routine reads three files containing the flow velocities (V), the water depths (D) and the substrate types (S), as derived from a hydrodynamic/hydraulic simulation of the river reach (each node of the computational grid being an array element) and three additional user-defined parameters (the inference process to be implemented, the management scenario and the defuzzification method) and calls the relevant subroutines to implement the selected inference process. The output of Habfuzz is a .txt file containing an array with the habitat suitabilities for each node of the initial computational grid of the hydrodynamic simulation and a .txt log file containing the internal parameters of the inference process.
 
 As described in detail by Ross (2010) and implemented in Habfuzz, the process of deriving the fuzzy-based habitat suitability includes three steps:  
 
