### 2016-10-12 #82 @chtheodoro
Dear all,

Habfuzz has been now tested and runs on Windows, Ubuntu and OS X. I have provided the relevant instructions in the README.md. Another subroutine was also added and now the tool uses Fuzzy logic or Fuzzy Bayesian inference depending on the user's selection.
@arfon I have changed the README.md to provide only the basic information, while indicating the reader/user to consult the Habfuzz manual for details. Is this ok with you? I also changed the paper.md and the paper.bib, how can I apply these changes to the submission of JOSS?
