+++ README.md
@@ -1,4 +1,6 @@
-# Habfuzz
+![onlinelogomaker-101416-1201-6434](https://cloud.githubusercontent.com/assets/21544603/19381943/a169dd32-9206-11e6-8207-055b982a6031.jpg)
+
+  
 Habfuzz is a FORTRAN tool, which calculates the instream hydraulic habitat suitability by utilizing  
 1.	Τhe Mamdani - Assilian fuzzy inference process (Mamdani and Assilian, 1975) and  
 2.	The Bayesian joint probability inference process as described in Brookes et al. (2010), with fuzzified inputs.
