### 2016-10-16 #82 @chtheodoro
@labarba and @arfon is there anything else I need to do in order for the review process to continue?
