### 2016-10-16 #82 @labarba
@fonnesbeck : Are you able to review this submission on Ubuntu or OS X ?
