### 2016-10-16 #82 @fonnesbeck
First, thanks for making the code cross-platform!

Here are couple of specific notes:

- misspelled "Brooks" on p.5 of manual
- `mcompile.sh` is not executable when pulled from GitHub
- the Geany text editor is cited as a dependency. Unless I am confused, there is no way for a text editor to be a dependency.

The only major issue I see with the package is that there does not appear to be any unit tests to verify that the constituent functions in Habfuzz behave as expected. While there are test cases provided, this is not the same as having tests that will fail when important bugs are introduced to the code (either by developers or users).
