### 2016-10-16 #82 @labarba
@fonnesbeck Did you install the software, then? If installation went without problems, do check the relevant items on the checklist at the top of this issue. Are you satisfied with the statement of need? See other check items above.
