### 2016-10-17 #82 @fonnesbeck
I was able to install and run the software. The outstanding issues, based on the review criteria include:

- `paper.md` seems to be incomplete. e.g. no references or author list. It appears much of the missing information could be moved from the README.
- no automated tests
