### 2016-10-17 3aa0638
Update paper.md
