+++ paper/paper.md
@@ -35,3 +35,11 @@ while the fuzzy Bayesian process includes (i) fuzzification of V and D, (ii) app
 ![figure_2](https://cloud.githubusercontent.com/assets/21544603/18706143/7f3e39fc-7ff9-11e6-8712-aaea188f8bc6.png)
 
 # References
+Brookes C.J., Kumar V. and Lane S.N. 2010. A comparison of Fuzzy, Bayesian and Weighted Average formulations of an in-stream habitat suitability model. Proceedings of the International Congress on Environmental Modelling and Software, 5-8 Jul 2010, Ottawa, Canada. Available at http://www.iemss.org/iemss2010/papers/S20/S.20.07.Model%20selection%20and%20uncertainty%20A%20comparison%20of%20Fuzzy,%20Bayesian%20and%20Weighted%20Average%20formulations%20of%20an%20instream%20habitat%20suitability%20model%20-%20CHRISTOPHER%20BROOKES.pdf
+
+Mamdani E.H. and Assilian S. 1975. An experiment in linguistic synthesis with a fuzzy logic
+Controller. International Journal of Man-Machine Studies 7: 1-13.
+
+Ross T.J. 2010. Fuzzy logic with engineering applications. Third Edition, John Wiley and Sons, UK.
+
+Zadeh L.A. 1965. Fuzzy sets. Information and Control 8: 338–353.
