### 2016-10-17 #82 @chtheodoro
Thank you @fonnesbeck for the quick reply!

I removed the 'Geany' text editor from the dependencies list. The paper.md is also updated with the reference list (the authors are already mentioned in the top box named 'authors'). Corrected 'Brookes' spelling also.

I am now working on the automated tests... I will let you know soon.
