### 2016-10-17 #82 @arfon
> I removed the 'Geany' text editor from the dependencies list. The paper.md is also updated with the reference list (the authors are already mentioned in the top box named 'authors'). Corrected 'Brookes' spelling also.

@chtheodoro - you don't need to include the references in the `References` section of `paper.md`. Instead, you just need to list them in the `paper.bib` and cite the references inline in your paper body. See this paper as an example: https://github.com/paulcon/active_subspaces/tree/master/JOSS_Paper
