+++ habfuzz/depths.txt
@@ -1,4 +1,11 @@
-3
-0.29
-0.08
-0.80
+10
+0.42
+0.37
+0.27
+0.17
+0.55
+0.04
+0.05
+0.09
+0.54
+0.76
