+++ habfuzz/substrates.txt
@@ -1,4 +1,11 @@
-3
-0.070
-0.050
+10
 0.040
+0.030
+0.070
+0.070
+0.030
+0.050
+0.070
+0.022
+0.030
+0.070
