+++ habfuzz/velocities.txt
@@ -1,4 +1,11 @@
-3
+10
 0.28
-0.05
-0.46
+0.29
+0.01
+0.19
+0.68
+1.15
+0.89
+0.07
+0.31
+0.23
