### 2016-10-19 291b53d
Update paper.md
