+++ paper/paper.bib
@@ -33,7 +33,7 @@
     url = {https://mitpress.mit.edu/books/fortran-95-handbook},
 }
 
-@online{github,
+@online{chtheodoro,
     author = {Theodoropoulos C. and Skoulikidis N. and Stamou A.},
     title = {Habfuzz: A Fortran tool to calculate the instream hydraulic habitat suitability based on fuzzy logic},
     year = 2016,
