### 2016-10-19 2da6f73
Update paper.md
