+++ paper/paper.bib
@@ -7,7 +7,7 @@
 }
 
 @conference{bayes,
-    author = {Brooks C.J. and Kumar V. and Lane S.N.},
+    author = {Brookes C.J. and Kumar V. and Lane S.N.},
     title = {A comparison of Fuzzy, Bayesian and Weighted Average formulations of an in-stream habitat suitability model},
     conference = {International Congress on Environmental Modelling and Software},
     year = {2010},
