### 2016-10-19 #82 @chtheodoro
Automated tests (unit tests) are now added in the tests folder of the program, along with a small README.md. I also updated the paper.md and paper.bib according to @arfon 's indications.

I hope I didn't miss anything... Please let me know if we are ok now.
Thank you!
