### 2016-10-21 #82 @chtheodoro
@labarba  and @fonnesbeck  when you are able we can continue with the review process. I applied all of @fonnesbeck 's indications.
