### 2016-10-21 #82 @fonnesbeck
Thanks. I will check the tests as soon as I can.
