+++ paper/paper.bib
@@ -35,7 +35,7 @@
 
 @online{chtheodoro,
     author = {Theodoropoulos C. and Skoulikidis N. and Stamou A.},
-    title = {Habfuzz: A Fortran tool to calculate the instream hydraulic habitat suitability based on fuzzy logic},
+    title = {Habfuzz: A tool to calculate the instream hydraulic habitat suitability based on fuzzy logic and fuzzy Bayesian inference},
     year = 2016,
     url = {https://github.com/chtheodoro/habfuzz},
     urldate = {2016-09-21}
