### 2016-10-21 #82 @fonnesbeck
@chtheodoro Unfortunately, I am having trouble installing the FRUIT testing framework. I have [filed a bug report](https://sourceforge.net/p/fortranxunit/bugs/5/), but given that the last report is 2 years old, I'm not sure how soon it will get a response. If you have an idea of where things are going wrong, please let me know (I am not much of a Ruby expert).

BTW, you state that Ruby/rake are optional dependencies for testing, but it seems that FRUIT requires them for installation, so they do not appear to be optional.
