### 2016-10-21 #82 @chtheodoro
I had the same trouble on Windows. However, without installing FRUIT, just put the fruit.95 file from my repository together with the test file, the test driver file, the fdeclarations.f95 and the relevant subroutine to test, compile with gfortran and you should be able to run. I tried it on a Mac and it works.
