### 2016-10-21 #82 @chtheodoro
If it runs only with fruit.f95 and nothing more please let me know in order to change the README file to make things easier...
