+++ tests/README.md
@@ -1,4 +1,9 @@
-Test application instructions:
+Each test has been named according to the relevant subroutine, which is to be tested.  
+For the 'fuzzifier' subroutine, the test calculates the sum of of the fuzzified inputs (separately for V, D and S) and giving a 'pass' if the sum equals 1.  
+For the 'rules' subroutine, the test calculates the total probability of all observations (VxDxSxHS) and giving a 'pass' if the sum of all probability observations equals 1.  
+All other tests, apply given input values and give a 'pass' when the output is the one expected by the test.
+
+To apply the tests:
 
 1. Put the test (e.g. test_centroid.f95), along with its relevant driver file (test_centroid_driver.f95) in a separate folder of your choice  
 2. Copy the files fdeclarations.f95 and the relevant subroutine (e.g. centroid.f95) from the 'habfuzz' subfolder and paste them to the previous folder  
