### 2016-10-23 #82 @chtheodoro
@fonnesbeck my apologies for bothering you about this, but did you manage to run the tests by using only the fruit.f95 file according to the updated README instructions?
