### 2016-10-23 #82 @fonnesbeck
Can you create a script to automate the testing? Moving individual files around is not ideal for users or developers. What is the issue with running the tests in place?
