### 2016-10-23 #82 @chtheodoro
Each test calls a subroutine to test and I thought that the subroutine file should be in the same directory with the test in order to compile. But I suppose I will find some way to automate the process...
