### 2016-10-23 #82 @chtheodoro
@fonnesbeck I found a solution to automate the test through makefiles but this would require one folder per test. Is it ok with you? This means 7 subfolders in the tests folder.
