### 2016-10-24 #82 @fonnesbeck
Yes, perhaps just a bash script to create a temporary directory, copy the files, run the test and clean up would be sufficient.
