### 2016-10-25 ba05e8c
Automated unit test
