+++ habfuzz/suitability.txt
@@ -1,10 +1,10 @@
-  0.671
-  0.440
-  0.690
-  0.737
+  0.660
+  0.500
+  0.500
+  0.700
   0.300
  -1.000
  -1.000
-  0.367
-  0.480
+  0.233
+  0.500
   0.500
