### 2016-10-25 #82 @chtheodoro
@fonnesbeck I made the unit tests fully automated by using makefiles. You only need to write a specific command (see the README of the tests folder) and the test runs. Tested on both linux and os x!!
