### 2016-10-25 #82 @fonnesbeck
That seems to work, thanks.

The only outstanding item I see is that there are no references (with DOI) at the bottom of `paper.md`.
