### 2016-10-25 #82 @arfon
> The only outstanding item I see is that there are no references (with DOI) at the bottom of `paper.md`.

@fonnesbeck - I need to compile the paper for you (this should pull in the references from `paper.bib`. I'll update the thread when I've done this.
