### 2016-10-25 #82 @arfon
@fonnesbeck here's the PDF: [10.21105.joss.00082.pdf](https://github.com/openjournals/joss-reviews/files/552019/10.21105.joss.00082.pdf)
