### 2016-10-25 0a7bebf
Update paper.md
