### 2016-10-26 #82 @fonnesbeck
Looks good to me. Thanks for your patience and diligence in addressing my comments.

@labarba Everything looks good to me now.
