### 2016-10-26 #82 @labarba
Thanks ever so much, @fonnesbeck, for your diligent review. And good job, @chtheodoro, with responding to the reviewer requests.

@arfon : This submission can now be accepted in JOSS. Yay!
