### 2016-10-26 #82 @arfon
> @arfon : This submission can now be accepted in JOSS. Yay!

✨  this is great!

@chtheodoro at this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
