+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @book{Fortran,
-    author = {Adams J.C. and Brainerd W.S. and Martin J.T. and Smith B.T. and Wagener J.L.},
+    author = {Adams, J.C. and Brainerd, W.S. and Martin, J.T. and Smith, B.T. and Wagener, J.L.},
     title = {Fortran 95 Handbook - Complete ISO/ANSI Reference},
     year = {1997},
     publisher = {The MIT Press, Cambridge, Massachusetts, London, England},
@@ -7,7 +7,7 @@
 }
 
 @conference{bayes,
-    author = {Brookes C.J. and Kumar V. and Lane S.N.},
+    author = {Brookes, C.J. and Kumar, V. and Lane, S.N.},
     title = {A comparison of Fuzzy, Bayesian and Weighted Average formulations of an in-stream habitat suitability model},
     conference = {International Congress on Environmental Modelling and Software},
     year = {2010},
@@ -16,7 +16,7 @@
 }
 
 @article{fis,
-    author = {Mamdani E.H. and Assilian S.},
+    author = {Mamdani, E.H. and Assilian, S.},
     title = {An experiment in linguistic synthesis with a fuzzy logic Controller},
     year = {1975},
     journal = {International Journal of Man-Machine Studies},
@@ -26,7 +26,7 @@
 }
 
 @book{fuzzy,
-    author = {Ross T.J.},
+    author = {Ross, T.J.},
     title = {Fuzzy logic with engineering applications},
     year = {2010},
     publisher = {John Wiley and Sons, UK},
@@ -34,7 +34,7 @@
 }
 
 @online{chtheodoro,
-    author = {Theodoropoulos C. and Skoulikidis N. and Stamou A.},
+    author = {Theodoropoulos, C. and Skoulikidis, N. and Stamou, A.},
     title = {Habfuzz: A tool to calculate the instream hydraulic habitat suitability based on fuzzy logic and fuzzy Bayesian inference},
     year = 2016,
     url = {https://github.com/chtheodoro/habfuzz},
