### 2016-10-26 #82 @chtheodoro
Thank you @fonnesbeck  and @labarba !!
@arfon the zenodo DOI is 10.5281/zenodo.163291.

I've notice that in the generated PDF, only the names are displayed in the references... Is there anything I did wrong in the paper.bib file? Normally, for example it should be (Mamdani and Assilian, 1975) but it is (E.H. and S., 1975).

And, not all authors have an ORCID ID. If it is necessary, please let me know to arrange this..
