### 2016-10-26 e3ce3de
Update paper.md
