### 2016-10-26 d8f6b4e
Update paper.md
