### 2016-10-26 2fd8489
Update paper.md
