### 2016-10-27 #82 @chtheodoro
@arfon I finally managed to merge your branch with the master and everything seems ok now. Please let me know if there is anything else I should do.
