### 2016-10-27 #82 @arfon
[10.21105.joss.00082.pdf](https://github.com/openjournals/joss-reviews/files/555583/10.21105.joss.00082.pdf)

> I've notice that in the generated PDF, only the names are displayed in the references... Is there anything I did wrong in the paper.bib file? Normally, for example it should be (Mamdani and Assilian, 1975) but it is (E.H. and S., 1975).

@chtheodoro is this looking better in the new compiled PDF above?
