### 2016-10-27 #82 @arfon
@fonnesbeck many thanks for the excellent review. @labarba ✨  for editing this one.

@chtheodoro your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00082 🚀 🎉 💥
